Поместила в пакет ru/mmtr, провела рефакторинг
