+++ src/main/java/ru/mmtr/dictionarySpring/Main.java
@@ -1,15 +1,15 @@
-package dictionarySpring;
+package ru.mmtr.dictionarySpring;
 
-import dictionarySpring.configuration.SpringConfig;
-import dictionarySpring.view.Console;
+import ru.mmtr.dictionarySpring.configuration.SpringConfig;
+import ru.mmtr.dictionarySpring.view.Console;
 import org.springframework.context.ApplicationContext;
 import org.springframework.context.annotation.AnnotationConfigApplicationContext;
 
 public class Main {
     public static void main(String[] args) {
         ApplicationContext context = new AnnotationConfigApplicationContext(SpringConfig.class);
-        Console test = context.getBean(Console.class);
-        test.menuSelectDictionary();
-        test.menuSelectActions();
+        Console console = context.getBean(Console.class);
+        console.menuSelectDictionary();
+        console.menuSelectActions();
     }
 }
