+++ src/main/java/ru/mmtr/dictionarySpring/exception/TypeNotFoundException.java
@@ -1,4 +1,4 @@
-package dictionarySpring.exception;
+package ru.mmtr.dictionarySpring.exception;
 
 public class TypeNotFoundException extends Exception{
     public TypeNotFoundException(String message){
