+++ src/main/java/ru/mmtr/dictionarySpring/view/Formation.java
@@ -1,6 +1,6 @@
-package dictionarySpring.view;
+package ru.mmtr.dictionarySpring.view;
 
-import dictionarySpring.model.DictionaryLine;
+import ru.mmtr.dictionarySpring.model.DictionaryLine;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
